+++ src/main/java/calculator/level2/ArithmeticCalculator.java
@@ -1,0 +1,44 @@
+package calculator.level2;
+
+public class ArithmeticCalculator extends Calculator {
+    private int firstNumber;
+    private int secondNumber;
+    private String operation;
+
+    public int getFirstNumber() {
+        return firstNumber;
+    }
+
+    public void setFirstNumber(int firstNumber) {
+        this.firstNumber = firstNumber;
+    }
+
+    public int getSecondNumber() {
+        return secondNumber;
+    }
+
+    public void setSecondNumber(int secondNumber) {
+        this.secondNumber = secondNumber;
+    }
+
+    public String getOperation() {
+        return operation;
+    }
+
+    public void setOperation(String operation) {
+        this.operation = operation;
+    }
+
+    @Override
+    public double calculate() {
+        double result = 0;
+        switch (operation) {
+            case "+": result = firstNumber + secondNumber; break;
+            case "-": result = firstNumber - secondNumber; break;
+            case "*": result = firstNumber * secondNumber; break;
+            case "/": result = firstNumber / secondNumber; break;
+        }
+        resultList.add(result);
+        return result;
+    }
+}
